refactor: move ECS stuff into cluster model file
